Initial #5

+++ ionion/src/main/java/gr/rege/ionion/info/DiskUsageInfo.java
@@ -2,16 +2,16 @@ package gr.rege.ionion.info;
 
 public class DiskUsageInfo 
 {
-	public int http_docs;
-	public int https_docs;
-	public int subdomains;
-	public int web_users;
-	public int anonftp;
-	public int logs;
-	public int dbases;
-	public int mailboxes;
-	public int webapps;
-	public int maillists;
-	public int domaindumps;
-	public int configs;
+	public long http_docs;
+	public long https_docs;
+	public long subdomains;
+	public long web_users;
+	public long anonftp;
+	public long logs;
+	public long dbases;
+	public long mailboxes;
+	public long webapps;
+	public long maillists;
+	public long domaindumps;
+	public long configs;
 }
